[ECR-94] fix: JpaRepository 타입 수정

+++ api/src/main/java/org/example/educheck/global/common/repository/ProducerFailedEventRepository.java
@@ -3,5 +3,5 @@ package org.example.educheck.global.common.repository;
 import org.example.educheck.global.common.entity.ProducerFailedEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProducerFailedEventRepository extends JpaRepository<ProducerFailedEvent, Integer> {
+public interface ProducerFailedEventRepository extends JpaRepository<ProducerFailedEvent, Long> {
 }
